fix: 잘못된 searchParam 타입 수정
